middleware, auth

+++ app/Http/Controllers/StoreController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\ProductCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class StoreController extends Controller
 {
@@ -24,6 +25,9 @@ class StoreController extends Controller
     }
 
     public function insert_product(Request $request){
+        if(!Gate::allows('insert-product')){
+            abort(403, 'Unauthorized action.');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'details' => 'nullable|string',
@@ -76,6 +80,9 @@ class StoreController extends Controller
     }
 
     public function update_product(Request $request, $product_id){
+        if(!Gate::allows('update-product')){
+            abort(403, 'Unauthorized action.');
+        }
         // Validate the incoming request data
         $request->validate([
             'name' => 'required|string|max:255',
@@ -122,6 +129,9 @@ class StoreController extends Controller
     }
 
     public function delete_product($product_id){
+        if(!Gate::allows('delete-product')){
+            abort(403, 'Unauthorized action.');
+        }
         $product = Product::findOrFail($product_id);
         $product->delete();
         return redirect()->route('store')->with('success', 'Product deleted successfully!');
